requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
Pillow>=9.0.0
scikit-image>=0.19.0
